Fan regulation
added fan regulation based on temperature and humidity

+++ src/sensors/dht11.h
@@ -35,8 +35,8 @@ enum dht11_status {
 
 struct dht11_reading {
     int status;
-    float temperature;
-    float humidity;
+    int temperature;
+    int humidity;
 };
 
 void DHT11_init(gpio_num_t);
