+++ src/sensors/sensors.h
@@ -1,7 +1,12 @@
 extern float BRIGHTNESS;
-extern float TEMPERATURE;
-extern float HUMIDITY;
+extern int TEMPERATURE;
+extern int HUMIDITY;
+//extern int HUMIDITY_SUM;
+//extern int AVERAGE_HUMIDITY;
 extern int WATER_DISTANCE;
+
+//Timer for measurements over a period of time
+//extern int timer_measurement = 0;
 
 // initialze sensors so sensors can read data
 void init_sensors();
